Fixed compiler warning about old-style casts
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@342 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ test/CreationTest/CreationTest.cpp
@@ -441,10 +441,10 @@ int main( int argc, char* argv[] )
         std::ostringstream oss;
         oss << "A" << zz;
 
-        writer.GetNamesTree()->AddValue( "TestDict", PdfString( oss.str() ), PdfVariant( (long)zz )  );
+        writer.GetNamesTree()->AddValue( "TestDict", PdfString( oss.str() ), PdfVariant( static_cast<long>(zz) )  );
     }
 
-    writer.GetNamesTree()->AddValue( "TestDict", PdfString( "Berta" ), PdfVariant( (long)42 )  );
+    writer.GetNamesTree()->AddValue( "TestDict", PdfString( "Berta" ), PdfVariant( 42L )  );
 
     printf("Setting document informations.\n\n");
     // Setup the document information dictionary
